+++ tests/test_recipes.py
@@ -7,10 +7,19 @@ import xarray as xr
 # need to import this way (rather than use pytest.lazy_fixture) to make it work with dask
 from pytest_lazyfixture import lazy_fixture
 
+from pangeo_forge_recipes.recipes.base import BaseRecipe
+
 all_recipes = [
     lazy_fixture("netCDFtoZarr_sequential_recipe"),
     lazy_fixture("netCDFtoZarr_sequential_multi_variable_recipe"),
 ]
+
+
+def test_to_pipelines_warns(netCDFtoZarr_sequential_recipe):
+    RecipeClass, file_pattern, kwargs, ds_expected, target = netCDFtoZarr_sequential_recipe
+    rec = RecipeClass(file_pattern, **kwargs)
+    with pytest.warns(FutureWarning):
+        rec.to_pipelines()
 
 
 @pytest.mark.parametrize("recipe_fixture", all_recipes)
@@ -22,6 +31,12 @@ def test_recipe(recipe_fixture, execute_recipe):
     execute_recipe(rec)
     ds_actual = xr.open_zarr(target.get_mapper()).load()
     xr.testing.assert_identical(ds_actual, ds_expected)
+
+    with rec.open_input(next(rec.iter_inputs())):
+        pass
+
+    with rec.open_chunk(next(rec.iter_chunks())):
+        pass
 
 
 @pytest.mark.parametrize("recipe_fixture", all_recipes)
@@ -170,3 +185,66 @@ def test_lock_timeout(netCDFtoZarr_sequential_recipe, execute_recipe):
     # if we're using a Dask executor.
     if execute_recipe.param in {"manual", "python", "prefect"}:
         assert p.call_args[1]["timeout"] == 1
+
+
+class MyRecipe(BaseRecipe):
+    def __init__(self) -> None:
+        super().__init__()
+        self.cache = {}
+        self.target = None
+        self.finalized = False
+        self.cache_inputs = True
+
+    @property
+    def prepare_target(self):
+        def _():
+            self.target = {}
+
+        return _
+
+    @property
+    def cache_input(self):
+        def _(input_key):
+            self.cache[input_key] = input_key
+
+        return _
+
+    @property
+    def store_chunk(self):
+        def _(chunk_key):
+            self.target[chunk_key] = self.cache[chunk_key]
+
+        return _
+
+    @property
+    def finalize_target(self):
+        def _():
+            self.finalized = True
+
+        return _
+
+    def iter_inputs(self):
+        return iter(range(4))
+
+    def iter_chunks(self):
+        return iter(range(4))
+
+
+def test_base_recipe():
+    recipe = MyRecipe()
+    recipe.to_function()()
+    assert recipe.finalized
+    assert recipe.target == {i: i for i in range(4)}
+
+    import dask
+
+    dask.config.set(scheduler="single-threaded")
+    recipe = MyRecipe()
+    recipe.to_dask().compute()
+    assert recipe.finalized
+    assert recipe.target == {i: i for i in range(4)}
+
+    recipe = MyRecipe()
+    recipe.to_prefect().run()
+    assert recipe.finalized
+    assert recipe.target == {i: i for i in range(4)}