+++ docs/development/release_notes.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## v0.4.0 - 2021-06-25
+
+- Fixed issue with recipe serialilzation causing high memory usage of Dask schedulers and workers when
+  executing recipes with Prefect or Dask (https://github.com/pangeo-forge/pangeo-forge-recipes/pull/160)[#160].
+- Added new methods `.to_dask()`, `to_prefect()`, and `.to_function()` for converting a recipe
+  to one of the Dask, Prefect, or Python execution plans. The previous method, `recpie.to_pipelines()`
+  is now deprecated.
+
 ## v0.3.4 - 2021-05-25
 
 - Added `copy_pruned` method to `XarrayZarrRecipe` to facilitate testing.